feat EleFile complete_flag and some refactorings

+++ partridge-mng-service/src/main/java/cn/nyaaar/partridgemngservice/controller/ElementController.java
@@ -50,7 +50,7 @@ public class ElementController {
     @Operation(summary = "分享 element", description = "通过主键id 分享 element")
     @GetMapping(value = "/share")
     @LogAnnotation
-    public R<String> shareElement(@RequestParam Integer elementId) {
+    public R<String> shareElement(@RequestParam Long elementId) {
         elementMngService.share(elementId);
         return new R<>();
     }
@@ -58,7 +58,7 @@ public class ElementController {
     @Operation(summary = "删除 element", description = "删除文件及其关联分片")
     @PostMapping(value = "/delete")
     @LogAnnotation
-    public R<String> delete(@RequestBody Integer eleId) {
+    public R<String> delete(@RequestBody Long eleId) {
         elementMngService.delete(eleId);
         return new R<>();
     }
